# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Route list ",
        "version": "1.0"
    },
    "paths": {
        "/api": {},
        "/api/v1/send_sms": {
            "post": {
                "tags": [
                    "Auth Api"
                ],
                "description": "use to send sms to user",
                "operationId": "8a6a727c1433d8e4ff86e112dd6813eb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "mobile": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/check_sms_code": {
            "post": {
                "tags": [
                    "Auth Api"
                ],
                "description": "use to check sms code that recieved by user",
                "operationId": "6776e1400fd13cb191c23b1b805576d3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "mobile": {
                                        "type": "string"
                                    },
                                    "code": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/register": {
            "post": {
                "tags": [
                    "Auth Api"
                ],
                "description": "use to signin user with recieved code",
                "operationId": "fd83cab1818154e58fe1ba7d34661c03",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "postal_code": {
                                        "type": "string"
                                    },
                                    "lat": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "lng": {
                                        "type": "number",
                                        "format": "double"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data saved",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/home": {
            "get": {
                "tags": [
                    "Home Page"
                ],
                "description": "get home page data",
                "operationId": "762138e69d864f137e7bcc4c732267af",
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/all_products": {
            "get": {
                "tags": [
                    "Products Page"
                ],
                "description": "get products page data",
                "operationId": "8a52a2bf2c81faf5db37666d41d02544",
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/newest_products": {
            "get": {
                "tags": [
                    "Products Page"
                ],
                "description": "get products page data",
                "operationId": "6f0d45fe830e165f2e88d31e040af146",
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/cheapest_products": {
            "get": {
                "tags": [
                    "Products Page"
                ],
                "description": "get products page data",
                "operationId": "e98bb029f58d92aad548ccfd7bf8808b",
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/most_expensive_products": {
            "get": {
                "tags": [
                    "Products Page"
                ],
                "description": "get products page data",
                "operationId": "4ab79235e9a951ef1ab58c7dfabb58f9",
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/most_viewed_products": {
            "get": {
                "tags": [
                    "Products Page"
                ],
                "description": "get products page data",
                "operationId": "ea06ffc556ddc74c7d842e2abe24484f",
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/products_by_category/{id}": {
            "get": {
                "tags": [
                    "Products Page"
                ],
                "description": "get products data by category id",
                "operationId": "91e5d2a4042bdf5b5a085b55d209cf78",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/products_by_brand/{id}": {
            "get": {
                "tags": [
                    "Products Page"
                ],
                "description": "get products data by brand id",
                "operationId": "effb338af4623ede6b0711f3fa28a04f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/search_product": {
            "post": {
                "tags": [
                    "Products Page"
                ],
                "description": "search product",
                "operationId": "ea7d13513536f8619780a90251848c9f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "search": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/product_details/{id}": {
            "get": {
                "tags": [
                    "Product Details"
                ],
                "description": "get product details data by product id",
                "operationId": "4f07ea0c3f4b174b0e7fa2ef8cc2d177",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "product id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/save_product_comment": {
            "post": {
                "tags": [
                    "Product Details"
                ],
                "description": "save user comment for product",
                "operationId": "866e6cee254de9df77dd5c99369c6c12",
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/profile": {
            "post": {
                "tags": [
                    "User info"
                ],
                "operationId": "29df1511df7eaf64ce674cacc25bfc50",
                "responses": {
                    "200": {
                        "description": "It's Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/user_addresses": {
            "post": {
                "tags": [
                    "User info"
                ],
                "operationId": "b5db4f6858b49ac58f5eeff1e2b4b009",
                "responses": {
                    "200": {
                        "description": "It's Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/user_received_orders": {
            "post": {
                "tags": [
                    "User info"
                ],
                "operationId": "d87f6e1cae6151cdfba1a5ddf96323d4",
                "responses": {
                    "200": {
                        "description": "It's Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/user_cancelled_orders": {
            "post": {
                "tags": [
                    "User info"
                ],
                "operationId": "235d4a1abfca4dbd613c6ae418f7091b",
                "responses": {
                    "200": {
                        "description": "It's Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/user_processing_orders": {
            "post": {
                "tags": [
                    "User info"
                ],
                "operationId": "865f8dc6b8fddf7d0d7c915722094ffd",
                "responses": {
                    "200": {
                        "description": "It's Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Bearer",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}